fix: adjust several import paths
